fix(mobile): verify quick actions on current app
