+++ apps/mobile/src/app/_layout.tsx
@@ -19,9 +19,9 @@ import { config } from "@/services/config";
 import { sendError } from "@/services/error-tracking";
 import { useGetInitialNotifications } from "@/services/linking";
 import { getExpoPostHog } from "@/services/observability";
-import { useQuickActions } from "@/services/quickActions";
-import { SceneName } from "@/types/SceneName";
+import { useQuickActions } from "@/services/quick-actions";
 import { store } from "@/store";
+import { SceneName } from "@/types/scene-name";
 
 // Wait for the assets to load before hiding the SplashScreen
 SplashScreen.preventAutoHideAsync()?.catch(sendError);
@@ -51,9 +51,7 @@ const App = () => {
 
   useTrackScreens();
   useGetInitialNotifications();
-  // Quick actions can be triggered from this unauthenticated-safe root
-  // mount, so navigation is gated on having resolved to the fully
-  // authenticated, onboarded route -- see `useQuickActions`'s docblock.
+  // Wait for authentication and onboarding before following a shortcut.
   useQuickActions(initialRouteName === SceneName.Swipe);
 
   // MAESTRO_E2E only: render magic modals inside the main window instead
